# frontend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Unified Holdings Tracker API",
    "version": "1.0.0",
    "description": "统一持仓追踪后端服务 API 文档",
    "contact": {
      "name": "API Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3001/api",
      "description": "本地开发服务器"
    }
  ],
...
  "paths": {
...
    "/portfolio": {
      "get": {
        "tags": ["Portfolio"],
        "summary": "获取投资组合列表",
        "description": "返回所有投资组合的基本信息",
        "responses": {
          "200": {
            "description": "成功返回组合列表",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": { "type": "string", "format": "uuid" },
                      "name": { "type": "string" },
                      "cash": { "type": "number" }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Portfolio"],
        "summary": "创建投资组合",
        "description": "创建新的投资组合",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "cash"],
                "properties": {
                  "name": { "type": "string", "description": "组合名称" },
                  "cash": { "type": "number", "description": "初始现金" },
                  "leverageInfo": {
                    "type": "object",
                    "properties": {
                      "totalCredit": {
                        "type": "number",
                        "description": "融资总额度"
                      },
                      "interestRate": { "type": "number", "description": "利率" }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "创建成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Portfolio" }
              }
            }
          },
          "400": {
            "description": "请求参数错误"
          }
        }
      }
    },
    "/portfolio/exchange-rates": {
      "get": {
        "tags": ["Currency"],
        "summary": "获取汇率",
        "description": "获取主要货币对人民币的实时汇率",
        "responses": {
          "200": {
            "description": "成功返回汇率",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "USD": { "type": "number", "nullable": true },
                    "HKD": { "type": "number", "nullable": true },
                    "CNY": { "type": "number", "example": 1.0 },
                    "updatedAt": { "type": "string", "format": "date-time" }
                  }
                }
              }
            }
          },
          "503": {
            "description": "汇率服务不可用"
          }
        }
      }
    },
    "/portfolio/{id}": {
      "get": {
        "tags": ["Portfolio"],
        "summary": "获取投资组合详情",
        "description": "获取指定投资组合的详细信息",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "成功返回组合详情",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PortfolioDetail" }
              }
            }
          },
          "404": {
            "description": "组合不存在"
          }
        }
      },
      "delete": {
        "tags": ["Portfolio"],
        "summary": "删除投资组合",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "204": {
            "description": "删除成功"
          },
          "404": {
            "description": "组合不存在"
          }
        }
      }
    },
    "/portfolio/{id}/transactions": {
      "get": {
        "tags": ["Transaction"],
        "summary": "获取交易记录",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "成功返回交易列表",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Transaction" }
                }
              }
            }
          },
          "404": {
            "description": "组合不存在"
          }
        }
      },
      "post": {
        "tags": ["Transaction"],
        "summary": "添加交易记录",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/TransactionInput" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "添加成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Transaction" }
              }
            }
          },
          "400": {
            "description": "请求参数错误"
          },
          "404": {
            "description": "组合不存在"
          }
        }
      }
    },
    "/portfolio/{id}/transactions/{txId}": {
      "delete": {
        "tags": ["Transaction"],
        "summary": "删除交易记录",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          },
          {
            "name": "txId",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "204": {
            "description": "删除成功"
          },
          "404": {
            "description": "组合或交易不存在"
          }
        }
      }
    },
    "/portfolio/{id}/stats": {
      "get": {
        "tags": ["Portfolio"],
        "summary": "获取投资组合统计信息",
        "description": "获取组合的详细统计数据，包括持仓、盈亏等",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          },
          {
            "name": "period",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["total", "daily", "weekly", "monthly", "yearly"]
            }
          },
          {
            "name": "startDate",
            "in": "query",
            "schema": { "type": "string", "format": "date" }
          },
          {
            "name": "endDate",
            "in": "query",
            "schema": { "type": "string", "format": "date" }
          }
        ],
        "responses": {
          "200": {
            "description": "成功返回统计信息",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PortfolioStats" }
              }
            }
          },
          "404": {
            "description": "组合不存在"
          }
        }
      }
    },
    "/portfolio/{id}/cash-recalc": {
      "get": {
        "tags": ["Portfolio"],
        "summary": "现金重算校验",
        "description": "重新计算组合的现金余额并返回差异",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "返回重算结果",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "portfolioId": { "type": "string" },
                    "name": { "type": "string" },
                    "currentCash": { "type": "number" },
                    "recalculatedCash": { "type": "number" },
                    "diff": { "type": "number" },
                    "steps": { "type": "array", "items": { "type": "object" } }
                  }
                }
              }
            }
          },
          "404": {
            "description": "组合不存在"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Portfolio": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "cash": { "type": "number" },
          "initialCash": { "type": "number" },
          "leverage": { "$ref": "#/components/schemas/LeverageInfo" },
          "transactions": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Transaction" }
          }
        }
      },
      "PortfolioDetail": {
        "allOf": [
          { "$ref": "#/components/schemas/Portfolio" },
          {
            "type": "object",
            "properties": {
              "positions": {
                "type": "array",
                "items": { "$ref": "#/components/schemas/Position" }
              },
              "totalMarketValue": { "type": "number" },
              "totalAssets": { "type": "number" },
              "netAssets": { "type": "number" },
              "netDepositedCash": { "type": "number" },
              "totalCommission": { "type": "number" },
              "leverageCost": { "type": "number" },
              "dailyPnl": { "type": "number" },
              "totalPnl": { "type": "number" }
            }
          }
        ]
      },
      "PortfolioStats": {
        "type": "object",
        "properties": {
          "portfolioId": { "type": "string" },
          "name": { "type": "string" },
          "cash": { "type": "number" },
          "leverage": { "$ref": "#/components/schemas/LeverageInfo" },
          "totalMarketValue": { "type": "number" },
          "totalAssets": { "type": "number" },
          "netAssets": { "type": "number" },
          "netDepositedCash": { "type": "number" },
          "totalCommission": { "type": "number" },
          "leverageCost": { "type": "number" },
          "totalDividendIncome": { "type": "number" },
          "dailyPnl": { "type": "number" },
          "totalPnl": { "type": "number" },
          "periodReturnPercent": { "type": "number", "nullable": true },
          "weeklyStats": { "type": "object" },
          "monthlyStats": { "type": "object" },
          "yearlyStats": { "type": "object" },
          "positions": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Position" }
          },
          "timestamp": { "type": "number" }
        }
      },
      "Transaction": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "date": { "type": "string", "format": "date-time" },
          "type": {
            "type": "string",
            "enum": [
              "BUY",
              "SELL",
              "DEPOSIT",
              "WITHDRAW",
              "DIVIDEND",
              "LEVERAGE_ADD",
              "LEVERAGE_REMOVE",
              "LEVERAGE_COST"
            ]
          },
          "assetCode": { "type": "string", "nullable": true },
          "quantity": { "type": "number", "nullable": true },
          "price": { "type": "number", "nullable": true },
          "amount": { "type": "number", "nullable": true },
          "commission": { "type": "number", "nullable": true },
          "leverageUsed": { "type": "number", "nullable": true },
          "notes": { "type": "string", "nullable": true }
        }
      },
      "TransactionInput": {
        "type": "object",
        "required": ["type", "date"],
        "properties": {
          "date": { "type": "string", "format": "date-time" },
          "type": {
            "type": "string",
            "enum": [
              "BUY",
              "SELL",
              "DEPOSIT",
              "WITHDRAW",
              "DIVIDEND",
              "LEVERAGE_ADD",
              "LEVERAGE_REMOVE",
              "LEVERAGE_COST"
            ]
          },
          "assetCode": { "type": "string" },
          "quantity": { "type": "number" },
          "price": { "type": "number" },
          "amount": { "type": "number" },
          "commission": { "type": "number" },
          "leverageUsed": { "type": "number" },
          "notes": { "type": "string" }
        }
      },
      "LeverageInfo": {
        "type": "object",
        "properties": {
          "totalAmount": { "type": "number" },
          "usedAmount": { "type": "number" },
          "availableAmount": { "type": "number" },
          "costRate": { "type": "number" }
        }
      },
...
    }
  }
}